requirements.txt
fastapi
spotipy
uvicorn
python-dotenv
yt_dlp
mutagen
